Major refactoring of EPOS Hardware Mediators: - Architecture and Machine type names are no longer typedefed: a single file chain is included and those files directly declare the final types. - Display and Keyboard continue to be handled multiply. - Multiple NIC architectures (e.g. Ethernet AND 15.4) on the same build are not supported. - Configs have been incorporated. - Templates dependent on Mediators have been eliminated.

+++ include/chronometer.h
@@ -64,7 +64,8 @@ public:
     void lap() { if(_start != 0) _stop = Alarm::_elapsed; }
     void stop() { lap(); }
 
-    Microsecond read() { return ticks() * 1000000 / frequency(); }
+    // The cast provides resolution for intermediate calculations
+    Microsecond read() { return static_cast<unsigned long long>(ticks()) * 1000000 / frequency(); }
 
 private:
     Time_Stamp ticks() {
